# fix benchmark fp8 blockwise group gemm