--- a/sgl-kernel/benchmark/bench_fp8_blockwise_group_gemm.py
+++ b/sgl-kernel/benchmark/bench_fp8_blockwise_group_gemm.py
@@ -97,9 +97,7 @@ def bench_deepgemm(
     )
 
     def run_deepgemm():
-        deep_gemm.m_grouped_gemm_fp8_fp8_bf16_nt_contiguous(
-            x_fp8, y_fp8, out, m_indices
-        )
+        deep_gemm.m_grouped_fp8_gemm_nt_contiguous(x_fp8, y_fp8, out, m_indices)
 
     # warmup
     for _ in range(num_warmup):